V20250823 Added Bootstrap Icons and More
Bootstrap icons are now used in the web app for actions that also are in the system tray.  Configs to allow for one or the other perhaps to come.  Code centralized in service for systray and server to call.   Next I will need to open web app main on startup and also handle the port being in use...that should be fun.