Jewish Static Public Domain Images added and MBC Mode
MBC Mode is used for Manchester Baptist Church.  Or if anyone else wants they can have a monthly change of a singular(per month) quote (Values Quotes).  It watched the month and when a new month comes it will change to the next value reverting to the first when all are done.  Simple rotation